TimeTableSpecification add lazy fetch, OrderTable add optional, movie dto & order dto added optional, polish and fixed converters with tests. statements added

+++ src/main/java/com/example/cinema/dto/CinemaHallDto.java
@@ -5,5 +5,5 @@ import com.example.cinema.core.SeatType;
 import java.util.EnumMap;
 import java.util.HashSet;
 
-public record CinemaHallDto(Long id, String name, Short seatsAmount, EnumMap<SeatType, HashSet<Integer>> seatsType) {
+public record CinemaHallDto(Long id, String name, Short seatsAmount, EnumMap<SeatType, HashSet<Short>> seatsType) {
 }
